gta4xl-common: mixer_gains_cs47l15: Set lower gains when using headsets
Have sensitive earphones with an inline mic, and less sensitive
headphones with none. Too much gain results in audible noise with the
sensitive earphones while needing to keep a lower volume. So, just
taking advantage of the headset and headphones paths to set a lower gain
for my more sensitive earphones and a higher one for my headphones.

Change-Id: I36bb9172050cc07e9738d6c21a45cf7b38b4d551